[REFACTOR] #419: 모델 관련 테스트 코드 도메인 패키지로 묶기
