+++ backend/src/test/java/autoever2/cartag/models/ModelControllerTest.java
@@ -1,4 +1,4 @@
-package autoever2.cartag.controller;
+package autoever2.cartag.models;
 
 import autoever2.cartag.models.dto.ModelDetailMappedDto;
 import autoever2.cartag.models.dto.ModelEfficiencyDataDto;
